Issue #21 Update methology of loading the track to play media player with low volume to obtain track duration.
This is required as the duration will not be populated (nor will the durationChanged
slot be called) if: media is paused instead of played, mediaplayer volume is set to 0
or mediaplayer is set to muted.
Therefore, this is the only way to be able to obtain the duration of the track.